Add PrettyPrinter That Returns Both Errors & Text
Add a new `pPrintText` function to the Brittany.Main module that behaves
similarly to the `coreIO` function, but instead only take a config &
input string & always returns a list of errors & either the unaltered or
the pretty printed text.

This new function will always check the generated text & ignores all
tracing and printing to stdout.

This a useful function for utility programs, like haskell-language-server,
that want to feed text to brittany & manipulate the output.